pyproject.toml
[project]
name = "04mcflsk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flasgger>=0.9.7.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.3.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "werkzeug>=3.1.3",
]
